Refactor step error tests
