+++ crates/rstest-bdd-macros/src/parsing/feature/test_support.rs
@@ -1,0 +1,147 @@
+//! Builders and helpers shared by feature parsing tests.
+
+use super::super::{ParsedStep, ScenarioData, extract_scenario_steps};
+use gherkin::{Background, LineCol, Scenario, Span, Step, StepType};
+
+#[expect(
+    unreachable_patterns,
+    reason = "StepType currently only has three variants"
+)]
+fn kw(ty: StepType) -> String {
+    match ty {
+        StepType::Given => "Given",
+        StepType::When => "When",
+        StepType::Then => "Then",
+        _ => unreachable!("kw() only supports Given, When, and Then"),
+    }
+    .to_string()
+}
+
+pub(crate) struct StepBuilder {
+    ty: StepType,
+    value: String,
+    docstring: Option<String>,
+    table: Option<gherkin::Table>,
+    keyword: Option<String>,
+}
+
+impl StepBuilder {
+    pub(crate) fn new(ty: StepType, value: &str) -> Self {
+        Self {
+            ty,
+            value: value.to_string(),
+            docstring: None,
+            table: None,
+            keyword: None,
+        }
+    }
+
+    pub(crate) fn with_keyword(mut self, kw: &str) -> Self {
+        self.keyword = Some(kw.to_string());
+        self
+    }
+
+    pub(crate) fn with_docstring(mut self, doc: &str) -> Self {
+        self.docstring = Some(doc.to_string());
+        self
+    }
+
+    pub(crate) fn with_table(mut self, rows: Vec<Vec<&str>>) -> Self {
+        self.table = Some(gherkin::Table {
+            rows: rows
+                .into_iter()
+                .map(|r| r.into_iter().map(str::to_string).collect())
+                .collect(),
+            span: Span { start: 0, end: 0 },
+            position: LineCol { line: 0, col: 0 },
+        });
+        self
+    }
+
+    pub(crate) fn build(self) -> Step {
+        Step {
+            keyword: self.keyword.unwrap_or_else(|| kw(self.ty)),
+            ty: self.ty,
+            value: self.value,
+            docstring: self.docstring,
+            table: self.table,
+            span: Span { start: 0, end: 0 },
+            position: LineCol { line: 0, col: 0 },
+        }
+    }
+}
+
+pub(crate) struct FeatureBuilder {
+    name: String,
+    background: Option<Vec<Step>>,
+    scenarios: Vec<(String, Vec<Step>)>,
+}
+
+impl FeatureBuilder {
+    pub(crate) fn new(name: &str) -> Self {
+        Self {
+            name: name.to_string(),
+            background: None,
+            scenarios: Vec::new(),
+        }
+    }
+
+    pub(crate) fn with_background(mut self, steps: Vec<Step>) -> Self {
+        self.background = Some(steps);
+        self
+    }
+
+    pub(crate) fn with_scenario(mut self, name: &str, steps: Vec<Step>) -> Self {
+        self.scenarios.push((name.to_string(), steps));
+        self
+    }
+
+    pub(crate) fn build(self) -> gherkin::Feature {
+        gherkin::Feature {
+            keyword: "Feature".into(),
+            name: self.name,
+            description: None,
+            background: self.background.map(|steps| Background {
+                keyword: "Background".into(),
+                name: String::new(),
+                description: None,
+                steps,
+                span: Span { start: 0, end: 0 },
+                position: LineCol { line: 0, col: 0 },
+            }),
+            scenarios: self
+                .scenarios
+                .into_iter()
+                .map(|(name, steps)| Scenario {
+                    keyword: "Scenario".into(),
+                    name,
+                    description: None,
+                    steps,
+                    examples: Vec::new(),
+                    tags: Vec::new(),
+                    span: Span { start: 0, end: 0 },
+                    position: LineCol { line: 0, col: 0 },
+                })
+                .collect(),
+            rules: Vec::new(),
+            tags: Vec::new(),
+            span: Span { start: 0, end: 0 },
+            position: LineCol { line: 0, col: 0 },
+            path: None,
+        }
+    }
+}
+
+pub(crate) fn assert_feature_extraction(
+    feature_builder: FeatureBuilder,
+    expected_steps: &[ParsedStep],
+    scenario_index: Option<usize>,
+) {
+    let feature = feature_builder.build();
+    let ScenarioData { steps, .. } = extract_scenario_steps(&feature, scenario_index)
+        .unwrap_or_else(|_| panic!("failed to extract scenario steps at index {scenario_index:?}"));
+    assert_eq!(
+        steps, expected_steps,
+        "extracted steps did not match expectation",
+    );
+}
